integrations/twitter-bot/requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
